- port from opensips r4526, credits goes to bogdan - 64bits compile warnings fixed
git-svn-id: https://openser.svn.sourceforge.net/svnroot/openser/trunk@4618 689a6050-402a-0410-94f2-e92a70836424

+++ modules_k/db_mysql/res.c
@@ -75,8 +75,8 @@ int db_mysql_get_columns(const db_con_t* _h, db_res_t* _r)
 			db_free_columns(_r);
 			return -4;
 		}
-		LM_DBG("allocate %d bytes for RES_NAMES[%d] at %p\n", sizeof(str), col,
-				RES_NAMES(_r)[col]);
+		LM_DBG("allocate %lu bytes for RES_NAMES[%d] at %p\n",
+				(unsigned long)sizeof(str), col, RES_NAMES(_r)[col]);
 
 		/* The pointer that is here returned is part of the result structure. */
 		RES_NAMES(_r)[col]->s = fields[col].name;
